requirements.txt
pandas==1.5.3
PyPDF2
aiohttp==3.8.3
aiosignal==1.3.1
alabaster==0.7.13
async-timeout==4.0.2
attrs==22.2.0
Babel==2.11.0
blobfile==2.0.1
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
dataclasses-json==0.5.7
docutils==0.19
faiss-cpu==1.7.3
filelock==3.9.0
Flask==2.2.2
frozenlist==1.3.3
greenlet==2.0.2
huggingface-hub==0.12.0
idna==3.4
imagesize==1.4.1
itsdangerous==2.1.2
Jinja2==3.1.2
lxml==4.9.2
MarkupSafe==2.1.2
marshmallow==3.19.0
marshmallow-enum==1.5.1
multidict==6.0.4
mypy-extensions==0.4.3
numpy==1.24.1
openai==0.26.4
packaging==23.0
pycryptodomex==3.17
pydantic==1.10.4
Pygments==2.14.0
python-dotenv==0.21.1
pytz==2022.7.1
PyYAML==6.0
regex==2022.10.31
requests==2.28.2
snowballstemmer==2.2.0
Sphinx==6.1.3
sphinxcontrib-applehelp==1.0.4
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
SQLAlchemy==1.4.46
tiktoken==0.1.2
tokenizers==0.13.2
tqdm==4.64.1
transformers==4.26.0
typing-inspect==0.8.0
typing_extensions==4.4.0
urllib3==1.26.14
Werkzeug==2.2.2
yarl==1.8.2
langchain==0.0.141